+++ Source/ImageProcessing.DomainModel/Model/Distributions/TwoParameterDistributions/UniformDistribution.cs
@@ -4,24 +4,24 @@ namespace ImageProcessing.Distributions.TwoParameterDistributions
 {
     public class UniformDistribution : IDistribution
     {
-        private double _a;
-        private double _b;
+        private decimal _a;
+        private decimal _b;
 
         public UniformDistribution() { }
-        public UniformDistribution(double a, double b)
+        public UniformDistribution(decimal a, decimal b)
         {
             _a = a;
             _b = b;
         }
 
-        public double FirstParameter => _a;
-        public double SecondParameter => _b;
+        public decimal FirstParameter => _a;
+        public decimal SecondParameter => _b;
 
-        public double GetMean() => (_a + _b) / 2;
-        public double GetVariance() => (_b - _a) * (_b - _a) / 12;
-        public double Quantile(double p) => _a + p * (_b - _a);
+        public decimal GetMean() => (_a + _b) / 2;
+        public decimal GetVariance() => (_b - _a) * (_b - _a) / 12;
+        public decimal Quantile(decimal p) => _a + p * (_b - _a);
 
-        public void SetParams((double, double) parms)
+        public void SetParams((decimal, decimal) parms)
         {
             _a = parms.Item1;
             _b = parms.Item2;
